refactor(web): deepen shared ttyd access seams
Keep shared-daemon behavior behind singular lifecycle, validated attach, remote tunnel, and browser-client module boundaries.
