+++ crates/rimz/src/mux/mod.rs
@@ -785,6 +785,8 @@ pub trait MuxBackend: Send + Sync {
     fn name(&self) -> MuxName;
     fn ensure_session(&self, opts: &SessionOptions) -> Result<()>;
     fn attach_command(&self, name: &str, config: &crate::config::MultiplexerConfig) -> CommandSpec;
+    /// Attach to a live session without creating a missing session.
+    fn attach_existing_command(&self, name: &str) -> CommandSpec;
     fn detach(&self, name: &str) -> Result<()>;
     /// Force-remove a session by name. A missing session is success — the goal
     /// state is "no session by that name", so callers can retire a stale or
