+++ crates/rimz/tests/integration/sidebar_launch.rs
@@ -192,6 +192,10 @@ impl MuxBackend for FakeBackend {
         CommandSpec::new("fake").arg(name)
     }
 
+    fn attach_existing_command(&self, name: &str) -> CommandSpec {
+        CommandSpec::new("fake").arg(name)
+    }
+
     fn detach(&self, _name: &str) -> rimz::mux::Result<()> {
         Ok(())
     }
